Process events before rendering
By handling events before rendering we make the UI more responsive
because they will take effect during current round of rendering. This
is not an issue when rendering in a loop, however, we render UI once
we send requests to every endpoint. Those times get significantly
longer as the number of endpoint increases.