feat: implement rate limiting
When 429 (Too Many Requests) or 503 (Service unavailable) status codes
are received, the fuzzer will try to resend the request after number of
seconds specified in `Retry-After` header. If the header is not present
it will use exponential backoff algorithm with a start value of 1
second.

If it won't succeed after `max_backoff` tries, it will abort fuzzing
current endpoint.

Closes: #23